+++ app/helpers/graph_link_helper.rb
@@ -1,0 +1,5 @@
+module GraphLinkHelper
+  def link_to_style(name, style)
+    link_to name, :style => style
+  end
+end
