+++ app/views/metrics/show.html.haml
@@ -7,11 +7,11 @@
 .graph-placeholder{:class => params[:style].to_s}
 
 %ul
-  %li= link_to "normal", :style => nil
-  %li= link_to "total by week", :style => "by_week"
-  %li= link_to "total by month", :style => "by_month"
-  %li= link_to "total by year", :style => "by_year"
-  %li= link_to "year vs year", :style => "year_v_year"
+  %li= link_to_style "normal", nil
+  %li= link_to_style "total by week", "by_week"
+  %li= link_to_style "total by month", "by_month"
+  %li= link_to_style "total by year", "by_year"
+  %li= link_to_style "year vs year", "year_v_year"
 
 
 %table.metrics
